% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAO_fishery.R
\name{calcFAO_fishery}
\alias{calcFAO_fishery}
\title{calcFAO_fishery}
\usage{
calcFAO_fishery(by_fishing_area = FALSE)
}
\arguments{
\item{by_fishing_area}{if TRUE, all fishing areas are provided. if FALSE, only marine and inlandwaters are distinguished}
}
\value{
Magpie object with fish data in dry matter
}
\description{
calculates fish data as Mt dry matter, distinguishing capture and aquaculture as well as fishing areas
}
\author{
Jasmin Wehner, Benjamin Leon Bodirsky
}
