% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFishCCimpacts.R
\name{calcFishCCimpacts}
\alias{calcFishCCimpacts}
\title{calcFishCCimpacts}
\usage{
calcFishCCimpacts(
  impacts = c("marine_capture", "marine_aquaculture"),
  total = FALSE,
  proxies = c("Mediterranean and Black Sea", "Pacific Antarctic")
)
}
\arguments{
\item{impacts}{fish types being affected by CC impacts. So far only available for "marine_capture" and "marine_aquaculture"}

\item{total}{if TRUE, aggregated to total production over all fish types}

\item{proxies}{fishing areas where we use proxie regions if no impact data is available.}
}
\value{
Data for each FAO Major Fishing area
}
\description{
Derive relative change of marine fish production (capture AND marine aquaculture) based on scenarios by Cheung et al 2018.
}
\examples{

\dontrun{ a <- calcOutput(type="FishCCimpacts")
}
}
\author{
Jasmin Wehner, Benjamin Leon Bodirsky
}
