% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAQUASTAT.R
\name{convertAQUASTAT}
\alias{convertAQUASTAT}
\title{convertAQUASTAT}
\usage{
convertAQUASTAT(x)
}
\arguments{
\item{x}{MAgPIE object containing AQUASTAT data on country level}
}
\value{
magpie objects with results on contury level
}
\description{
Convert data based on AQUASTAT database (http://www.fao.org/nr/water/aquastat/data/query/index.html?lang=en)
}
\examples{

\dontrun{ 
 readSource("AQUASTAT", subtype="ConsAgri", convert=TRUE)
}
}
\author{
Kristine Karstens
}
