% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAO_fishery.R
\name{convertFAO_fishery}
\alias{convertFAO_fishery}
\title{convertFAO_fishery}
\usage{
convertFAO_fishery(x, subtype)
}
\arguments{
\item{x}{magpie object with uncoverted source data}

\item{subtype}{"capture" takes all fishdata into account that has been declared as capture fishery
"aquaculture" takes all fishdata into account that has been listed as aquaculture fishery}
}
\value{
Fishery data as complete MAgPIE object on country level
}
\description{
Converts readFAO_fishery output to complete MAgPIE object containing fishery data on country level (in tonnes)
}
\seealso{
\code{\link{readSource}}
}
\author{
Jasmin Wehner,Benjamin Leon Bodirsky
}
