% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCheung2016.R
\name{correctCheung2016}
\alias{correctCheung2016}
\title{Correct FAO climate impact data from Cheung et al 2016}
\usage{
correctCheung2016(subtype)
}
\arguments{
\item{subtype}{"Area" data subtype. Areas in square km for each Large Marine Ecosystem obtained from Seaaroundus.org
"PrimProdinmgCday" data subtype. Primary Production in mg C day^-1 for each Large Marine Ecosystem obtained from Seaaroundus.org
"Degrees" data subtype. relative change to fishery production in degree1p5,degree2p5,degree3p5 scenarios. obtained from Cheung et al 2016}
}
\value{
magpie object of the corrected Cheung et al 2016 data and converted x_PrimProdinmgCday into tCyr^(-1)^km2^(-1)
}
\description{
Correct magpie objects
}
\seealso{
\code{\link{readSource}}
}
\author{
Jasmin Wehner
}
