% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctCheung2018.R
\name{correctCheung2018}
\alias{correctCheung2018}
\title{correctCheung2018}
\usage{
correctCheung2018(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from the read function.}

\item{subtype}{"General" data subtype. Areas in square km and Primary Production in mg C day^-1 or Mt yr^-1 for each Exclusive Economic Zone obtained from Seaaroundus.org
"models" data subtype: DBEM Model output for RCP2.6;RCP8.5 obtained from Cheung et al 2018
"ModelOutputDynModel" data subtype. Dynamic Model output for RCP2.6;RCP8.5 obtained from Cheung et al 2018}
}
\value{
Fishery data as complete MAgPIE object on country level
}
\description{
Converts readCheung2018 output to complete MAgPIE object containing fishery data on aggregated FAO Major Fishing areas
}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Leon Bodirsky, Jasmin Wehner
}
