% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfish-package.R
\docType{package}
\name{mrfish-package}
\alias{mrfish}
\alias{mrfish-package}
\title{mrfish: madrat data preparation for data connected to fish}
\description{
Package contains routines to prepare data for validation exercises.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrfish}
  \item Report bugs at \url{https://github.com/pik-piam/mrfish/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Leon Bodirsky \email{bodirsky@pik-potsdam.de}

Authors:
\itemize{
  \item Jasmin Wehner
}

}
