% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAQUASTAT.R
\name{readAQUASTAT}
\alias{readAQUASTAT}
\title{readAQUASTAT}
\usage{
readAQUASTAT(subtype = "ConsAgri")
}
\arguments{
\item{subtype}{"Conservation_agriculture_area_4454": multicropping factor on cropped (excluding fallow) land, 
\itemize{ 
\item \code{aquastat}:      4454|Conservation agriculture area (1000 ha) 
                            (4454_conservation_agriculture_area_in_1000_ha.csv)
\item \code{aquastatShare}: 4455|Commoditiy Balance LivestockConservation agriculture area as % of arable land area (%) 
                            (4455_conservation_agriculture_area_as_share_of_arable_land_areas.csv)
}}
}
\value{
magpie objects with results on contury level
}
\description{
Read in data based on AQUASTAT database (http://www.fao.org/nr/water/aquastat/data/query/index.html?lang=en)
}
\examples{

\dontrun{ 
 readSource("AQUASTAT", subtype="ConsAgri", convert=TRUE)
}

}
\author{
Kristine Karstens
}
