% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCheung2016.R
\name{readCheung2016}
\alias{readCheung2016}
\title{Read climate impact data on  marine fishery production for from Cheung 2016}
\usage{
readCheung2016(subtype)
}
\arguments{
\item{subtype}{"Area" data subtype. Areas in square km for each Large Marine Ecosystem obtained from Seaaroundus.org
"PrimProdinmgCday" data subtype. Primary Production in mg C day^-1 for each Large Marine Ecosystem obtained from Seaaroundus.org
"Degrees" data subtype. relative change to fishery production in degree1p5,degree2p5,degree3p5 scenarios. obtained from Cheung et al 2016}
}
\value{
magpie object of the fishery data with Area, Primary Productivity and Degrees subtypes
}
\description{
Read-in a csv file as magclass object
in absolut and percentage values
}
\seealso{
\code{\link{readSource}}
}
\author{
Jasmin Wehner
}
