% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCheung2018.R
\name{readCheung2018}
\alias{readCheung2018}
\title{Read climate impact data on  marine fishery production for from Cheung 2018}
\usage{
readCheung2018(subtype)
}
\arguments{
\item{subtype}{"General" data subtype. Areas in square km and Primary Production in mg C day^-1 for each Exclusive Economic Zone obtained from Seaaroundus.org
"models" data subtype. DBEM Model output for RCP2.6;RCP8.5 and  Dynamic Model output for RCP2.6;RCP8.5 obtained from Cheung et al 2018}
}
\value{
magpie object of the fishery data with respecitive model outputs
}
\description{
Read-in a csv file as magclass object
in percentage values
}
\seealso{
\code{\link{readSource}}
}
\author{
Jasmin Wehner
}
