% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAO_fishery.R
\name{readFAO_fishery}
\alias{readFAO_fishery}
\title{Read FAO fishery data}
\usage{
readFAO_fishery(subtype)
}
\arguments{
\item{subtype}{"capture" takes all fishdata into account that has been declared as capture fishery
"aquaculture" takes all data into account that has been listed as aquaculture fishery}
}
\value{
magpie object of the fishery data with Capture or Aquaculture
}
\description{
Read-in a csv file as magclass object
in tonnes
}
\seealso{
\code{\link{readSource}}
}
\author{
Benjamin Leon Bodirsky, Jasmin Wehner
}
